% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hourly_n105}
\alias{hourly_n105}
\title{Hourly stage data at station N105}
\format{
\subsection{\code{hourly_n105}}{

A data frame with 176034 rows and 2 columns:
\describe{
\item{datetime}{Time of observation}
\item{H}{Stage level (m)}
}
}
}
\source{
ARPA FVG -- \href{https://www.meteo.fvg.it/}{OSMER}
}
\usage{
hourly_n105
}
\description{
Observed hourly stage data 2003-2023 of station N105 provided
by ARPA FVG (OSMER)
}
\keyword{datasets}
