% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obs_n105}
\alias{obs_n105}
\title{Observed data of station N105}
\format{
\subsection{\code{obs_n105}}{

A data frame with 29 rows and 4 columns:
\describe{
\item{TimeStart}{Begin of sampling time}
\item{StationCode}{ID of the station}
\item{H}{Stage level (cm)}
\item{Q}{River discharge (m3/s)}
}
}
}
\source{
Regione Autonoma Friuli Venezia Giulia -- \href{https://www.regione.fvg.it/rafvg/cms/RAFVG/ambiente-territorio/tutela-ambiente-gestione-risorse-naturali/FOGLIA202/}{Servizio Gestione Risorse Idriche}
}
\usage{
obs_n105
}
\description{
Observed discharge and stage data 2008-2023 of station N105 provided
by Autonomous Region FVG (Servizio Gestione Risorse Idriche)
}
\keyword{datasets}
