% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorialhydrology-package.R
\docType{package}
\name{tutorialhydrology-package}
\alias{tutorialhydrology}
\alias{tutorialhydrology-package}
\title{tutorialhydrology: Tutorials for the 6th Workshop on Water Resources in Developing Countries}
\description{
Tutorials for the 6th Workshop on Water Resources in Developing Countries.
}
\author{
\strong{Maintainer}: Giovanni Bonafe \email{giovanni.bonafe@arpa.fvg.it} (\href{https://orcid.org/0000-0002-7815-8866}{ORCID})

}
\keyword{internal}
