% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obs_discharge}
\alias{obs_discharge}
\title{Observed discharge data}
\format{
\subsection{\code{obs_discharge}}{

A data frame with 9768 rows and 20 columns:
\describe{
\item{Variable}{Variable}
\item{Unit}{Unit}
\item{Value}{Value}
\item{IdMeasure}{ID of the measure}
\item{Date}{Day of sampling}
\item{TimeStart}{Begin of sampling time}
\item{TimeEnd}{End of sampling time}
\item{NoBranches}{Number of river's branches at the time of the measure}
\item{Instrument}{Instrument}
\item{Method}{Method}
\item{NotesMeasure}{Notes to the measure}
\item{StationCode}{ID of the station}
\item{Municipality}{Municipality}
\item{Province}{Province}
\item{Location}{Location}
\item{River}{River}
\item{Basin}{Basin}
\item{SubBasin}{Sub-basin}
\item{Description}{ID of the station or something else}
\item{NotesStation}{Notes to the station}
}
}
}
\source{
Regione Autonoma Friuli Venezia Giulia -- \href{https://www.regione.fvg.it/rafvg/cms/RAFVG/ambiente-territorio/tutela-ambiente-gestione-risorse-naturali/FOGLIA202/}{Servizio Gestione Risorse Idriche}
}
\usage{
obs_discharge
}
\description{
Observed discharge, stage, temperature and flow speed data 2004-2024 provided
by Autonomous Region FVG (Servizio Gestione Risorse Idriche)
}
\keyword{datasets}
